split up interface file into creator and verifier files
