+++ spv/creator_test.go
@@ -1,0 +1,760 @@
+package spv_test
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"testing"
+
+	"github.com/libsv/go-bc"
+	"github.com/libsv/go-bc/spv"
+	"github.com/libsv/go-bt/v2"
+	"github.com/stretchr/testify/assert"
+)
+
+type mockTxMerkleGetter struct {
+	txStoreFunc func(context.Context, string) (*bt.Tx, error)
+	mpStoreFunc func(context.Context, string) (*bc.MerkleProof, error)
+}
+
+func (m *mockTxMerkleGetter) Tx(ctx context.Context, txID string) (*bt.Tx, error) {
+	if m.txStoreFunc == nil {
+		return nil, errors.New("txGetterFunc in test is undefined")
+	}
+
+	return m.txStoreFunc(ctx, txID)
+}
+
+func (m *mockTxMerkleGetter) MerkleProof(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+	if m.txStoreFunc == nil {
+		return nil, errors.New("mpGetterFunc in test is undefined")
+	}
+
+	return m.mpStoreFunc(ctx, txID)
+}
+
+func TestSPVEnvelope_CreateEnvelope(t *testing.T) {
+	tests := map[string]struct {
+		tx     string
+		txFunc func(context.Context, string) (*bt.Tx, error)
+		mpFunc func(context.Context, string) (*bc.MerkleProof, error)
+		exp    spv.Envelope
+		expErr error
+	}{
+		"valid envelope created": {
+			tx: "0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				return nil, fmt.Errorf("this test should be be using txFunc, tx %s", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				mp, ok := map[string]*bc.MerkleProof{
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+						Index:  4,
+						TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"*",
+							"*",
+							"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+						},
+					},
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						Index:  6,
+						TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+						Nodes: []string{
+							"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+							"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+							"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			exp: spv.Envelope{
+				TxID:  "bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84",
+				RawTx: "0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000",
+				Parents: map[string]*spv.Envelope{
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						TxID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Proof: &bc.MerkleProof{
+							Index:  1,
+							TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+							Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+							Nodes: []string{
+								"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+								"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+								"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+							},
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						TxID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Proof: &bc.MerkleProof{
+							Index:  2,
+							TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+							Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+							Nodes: []string{
+								"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+								"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+								"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+							},
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+						TxID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+						Proof: &bc.MerkleProof{
+							Index:  4,
+							TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+							Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+							Nodes: []string{
+								"*",
+								"*",
+								"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+							},
+						},
+					},
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						TxID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Proof: &bc.MerkleProof{
+							Index:  6,
+							TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+							Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+							Nodes: []string{
+								"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+								"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+								"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+							},
+						},
+					},
+				},
+			},
+		},
+		"creating an envelope with tx that doesn't exist errors": {
+			tx: "0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				if txID == "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3" {
+					return nil, nil
+				}
+				return nil, fmt.Errorf("tx %s not defined for this test", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				mp, ok := map[string]*bc.MerkleProof{
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": nil,
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						Index:  6,
+						TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+						Nodes: []string{
+							"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+							"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+							"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			expErr: errors.New("could not find tx 9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3"),
+		},
+		"error when getting tx is handled": {
+			tx: "0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				if txID == "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34" {
+					return nil, errors.New("big bad error")
+				}
+				return nil, fmt.Errorf("tx %s not defined in this test", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				mp, ok := map[string]*bc.MerkleProof{
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": nil,
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+						Index:  4,
+						TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"*",
+							"*",
+							"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			expErr: errors.New("failed to get tx 5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34: big bad error"),
+		},
+		"error when getting merkle proof is handled": {
+			tx: "0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				return nil, fmt.Errorf("this test should be be using txFunc, tx %s", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				if txID == "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c" {
+					return nil, errors.New("bigger badder error")
+				}
+
+				mp, ok := map[string]*bc.MerkleProof{
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			expErr: errors.New("failed to get merkle proof for tx c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c: bigger badder error"),
+		},
+		"envelope needing multiple layers can be built": {
+			tx: "0200000002bcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4010000006b483045022100d18b422f8cc7c14444aa091fddcb2cf2276c3f7cf496fc186327366f72e8b03802204b6f1b2d2ca44c8a56766f287c951a14302992f55a3063b7fd5c85ec6f2c4f2641210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff6111fe484db08f2a3b6035e21423dc0a8251904951182a9f5b1c3165be79dade010000006b4830450221008a277ad76dfeb69dfdcb15b22ed4539e06d263710e527b676915e90341c19bee02205ae03bbcf4281c2d1470b9e60847f13631c9285557c94df07bc8b447370644784121021f163f44d261868142986872ea90f3155a83775fb9c00b4f3c517e95b30d3b3cfeffffff0200ca9a3b000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac1ed2f505000000001976a914e13aac4a3f01fd3154cc66fc9e285a33bd84d7df88ac6e000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				switch txID {
+				case "deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161": //nolint:goconst
+					return bt.NewTxFromString("02000000020a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb000000006b483045022100ed01b49050b67634557cf01ad1abcb7271886e65c2edb09c325df3804270ed6402203b86581c0305778e9fb9f4a6b9b1e46a3084e6b85f5676bbdabe31c8128308d5412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff0a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb010000006a4730440220085479effe11ea67ed66622b1b231ec250131cad773bd0b36e79ea81701e7e55022013f9c6ed94ce7099fe807cbf04095cd545b651b6bbe31645ff5ab143b26faf26412102b552f6dff6e03a36ec58d44b71e20d779e92dbe53b52ac08d8236bf87f6ed14dfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac94d3f505000000001976a9141dcd2b24fe8457c775e1a6280bf91d68f48186e988ac6d000000")
+				case "bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": //nolint:goconst
+					return bt.NewTxFromString("0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000")
+				case "eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a": //nolint:goconst
+					return bt.NewTxFromString("0200000003923b8561e34c743ec4e736973f99b4d99d6a6a779d85a29a1f900af435467257010000006b483045022100b3bb6382590f35ce3eeccf3e4e6cc575b2d0f47a5cfb333c77df4ef276cbaa8c022059040d6ce37292f10d7b70feb96d28784d701ca12d0d26adf149400c47e77128412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffbcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4000000006a473044022077eca6d671d07a19c622cf88cb421787e5a11ba3a4a537e00c74be8491966b84022035279d55889a163b53d029e8e65f3430756a59293208cdf371692bc1120276c6412103c2db1771958c949a26ba221dd048dd387f95ea5d761805bbb8c6cd2d1ed22a6afeffffff844ce271fb05ce083807cc48dc418d35604f2319946346c92d0705fb6bc21ebd000000006a47304402207a1814eefee55e9ee17653b7122cf6352be8d204ce51ecc9b92efa628295e24202200be4daba99d3a01681d74049eca3f256e523ad9d5bb71718409add4b3961417e4121038eae2a1e8fcf7f43788366ce0c5a39753e14bd561b92121306071abf5f5b34bcfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac0ad5f505000000001976a914e361cafbc05a5596fe4378879230000aa28990b088ac6c000000")
+				}
+				return nil, fmt.Errorf("tx %s not defined for test", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				mp, ok := map[string]*bc.MerkleProof{
+					"3f92e4567741afebdb26296c35c370e274d1b1455faaed2e27af499e6c65e759": nil,
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161": nil,
+					"eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a": nil,
+					"d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc": {
+						Index:  3,
+						TxOrID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+						Target: "67641544b6fe809a0037c9683df0983e973c77a7ed1e8eff3b0f8c9b74cd3472",
+						Nodes: []string{
+							"a6d45ccb68a19c05346d3412f78d154d5834380413cbe740ec3550469b487f6e",
+						},
+					},
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+						Index:  4,
+						TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"*",
+							"*",
+							"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+						},
+					},
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						Index:  6,
+						TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+						Nodes: []string{
+							"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+							"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+							"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+						},
+					},
+					"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92": {
+						Index:  3,
+						TxOrID: "57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			exp: spv.Envelope{
+				TxID:  "3f92e4567741afebdb26296c35c370e274d1b1455faaed2e27af499e6c65e759",
+				RawTx: "0200000002bcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4010000006b483045022100d18b422f8cc7c14444aa091fddcb2cf2276c3f7cf496fc186327366f72e8b03802204b6f1b2d2ca44c8a56766f287c951a14302992f55a3063b7fd5c85ec6f2c4f2641210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff6111fe484db08f2a3b6035e21423dc0a8251904951182a9f5b1c3165be79dade010000006b4830450221008a277ad76dfeb69dfdcb15b22ed4539e06d263710e527b676915e90341c19bee02205ae03bbcf4281c2d1470b9e60847f13631c9285557c94df07bc8b447370644784121021f163f44d261868142986872ea90f3155a83775fb9c00b4f3c517e95b30d3b3cfeffffff0200ca9a3b000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac1ed2f505000000001976a914e13aac4a3f01fd3154cc66fc9e285a33bd84d7df88ac6e000000",
+				Parents: map[string]*spv.Envelope{
+					"d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc": {
+						TxID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+						Proof: &bc.MerkleProof{
+							Index:  3,
+							TxOrID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+							Target: "67641544b6fe809a0037c9683df0983e973c77a7ed1e8eff3b0f8c9b74cd3472",
+							Nodes: []string{
+								"a6d45ccb68a19c05346d3412f78d154d5834380413cbe740ec3550469b487f6e",
+							},
+						},
+					},
+					"deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161": {
+						TxID:  "deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161",
+						RawTx: "02000000020a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb000000006b483045022100ed01b49050b67634557cf01ad1abcb7271886e65c2edb09c325df3804270ed6402203b86581c0305778e9fb9f4a6b9b1e46a3084e6b85f5676bbdabe31c8128308d5412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff0a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb010000006a4730440220085479effe11ea67ed66622b1b231ec250131cad773bd0b36e79ea81701e7e55022013f9c6ed94ce7099fe807cbf04095cd545b651b6bbe31645ff5ab143b26faf26412102b552f6dff6e03a36ec58d44b71e20d779e92dbe53b52ac08d8236bf87f6ed14dfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac94d3f505000000001976a9141dcd2b24fe8457c775e1a6280bf91d68f48186e988ac6d000000",
+						Parents: map[string]*spv.Envelope{
+							"eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a": {
+								TxID:  "eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a",
+								RawTx: "0200000003923b8561e34c743ec4e736973f99b4d99d6a6a779d85a29a1f900af435467257010000006b483045022100b3bb6382590f35ce3eeccf3e4e6cc575b2d0f47a5cfb333c77df4ef276cbaa8c022059040d6ce37292f10d7b70feb96d28784d701ca12d0d26adf149400c47e77128412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffbcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4000000006a473044022077eca6d671d07a19c622cf88cb421787e5a11ba3a4a537e00c74be8491966b84022035279d55889a163b53d029e8e65f3430756a59293208cdf371692bc1120276c6412103c2db1771958c949a26ba221dd048dd387f95ea5d761805bbb8c6cd2d1ed22a6afeffffff844ce271fb05ce083807cc48dc418d35604f2319946346c92d0705fb6bc21ebd000000006a47304402207a1814eefee55e9ee17653b7122cf6352be8d204ce51ecc9b92efa628295e24202200be4daba99d3a01681d74049eca3f256e523ad9d5bb71718409add4b3961417e4121038eae2a1e8fcf7f43788366ce0c5a39753e14bd561b92121306071abf5f5b34bcfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac0ad5f505000000001976a914e361cafbc05a5596fe4378879230000aa28990b088ac6c000000",
+								Parents: map[string]*spv.Envelope{
+									"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92": {
+										TxID: "57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+										Proof: &bc.MerkleProof{
+											Index:  3,
+											TxOrID: "57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+											Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+											Nodes: []string{
+												"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+												"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+												"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+											},
+										},
+									},
+									"d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc": {
+										TxID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+										Proof: &bc.MerkleProof{
+											Index:  3,
+											TxOrID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+											Target: "67641544b6fe809a0037c9683df0983e973c77a7ed1e8eff3b0f8c9b74cd3472",
+											Nodes: []string{
+												"a6d45ccb68a19c05346d3412f78d154d5834380413cbe740ec3550469b487f6e",
+											},
+										},
+									},
+									"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": {
+										TxID:  "bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84",
+										RawTx: "0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000",
+										Parents: map[string]*spv.Envelope{
+											"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+												TxID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+												Proof: &bc.MerkleProof{
+													Index:  1,
+													TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+													Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+													Nodes: []string{
+														"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+														"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+														"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+													},
+												},
+											},
+											"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+												TxID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+												Proof: &bc.MerkleProof{
+													Index:  2,
+													TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+													Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+													Nodes: []string{
+														"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+														"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+														"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+													},
+												},
+											},
+											"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+												TxID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+												Proof: &bc.MerkleProof{
+													Index:  4,
+													TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+													Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+													Nodes: []string{
+														"*",
+														"*",
+														"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+													},
+												},
+											},
+											"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+												TxID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+												Proof: &bc.MerkleProof{
+													Index:  6,
+													TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+													Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+													Nodes: []string{
+														"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+														"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+														"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+													},
+												},
+											},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		"missing tx multiple layers down causes error": {
+			tx: "0200000002bcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4010000006b483045022100d18b422f8cc7c14444aa091fddcb2cf2276c3f7cf496fc186327366f72e8b03802204b6f1b2d2ca44c8a56766f287c951a14302992f55a3063b7fd5c85ec6f2c4f2641210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff6111fe484db08f2a3b6035e21423dc0a8251904951182a9f5b1c3165be79dade010000006b4830450221008a277ad76dfeb69dfdcb15b22ed4539e06d263710e527b676915e90341c19bee02205ae03bbcf4281c2d1470b9e60847f13631c9285557c94df07bc8b447370644784121021f163f44d261868142986872ea90f3155a83775fb9c00b4f3c517e95b30d3b3cfeffffff0200ca9a3b000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac1ed2f505000000001976a914e13aac4a3f01fd3154cc66fc9e285a33bd84d7df88ac6e000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				switch txID {
+				case "deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161":
+					return bt.NewTxFromString("02000000020a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb000000006b483045022100ed01b49050b67634557cf01ad1abcb7271886e65c2edb09c325df3804270ed6402203b86581c0305778e9fb9f4a6b9b1e46a3084e6b85f5676bbdabe31c8128308d5412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff0a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb010000006a4730440220085479effe11ea67ed66622b1b231ec250131cad773bd0b36e79ea81701e7e55022013f9c6ed94ce7099fe807cbf04095cd545b651b6bbe31645ff5ab143b26faf26412102b552f6dff6e03a36ec58d44b71e20d779e92dbe53b52ac08d8236bf87f6ed14dfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac94d3f505000000001976a9141dcd2b24fe8457c775e1a6280bf91d68f48186e988ac6d000000")
+				case "bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84":
+					return bt.NewTxFromString("0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000")
+				case "eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a":
+					return bt.NewTxFromString("0200000003923b8561e34c743ec4e736973f99b4d99d6a6a779d85a29a1f900af435467257010000006b483045022100b3bb6382590f35ce3eeccf3e4e6cc575b2d0f47a5cfb333c77df4ef276cbaa8c022059040d6ce37292f10d7b70feb96d28784d701ca12d0d26adf149400c47e77128412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffbcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4000000006a473044022077eca6d671d07a19c622cf88cb421787e5a11ba3a4a537e00c74be8491966b84022035279d55889a163b53d029e8e65f3430756a59293208cdf371692bc1120276c6412103c2db1771958c949a26ba221dd048dd387f95ea5d761805bbb8c6cd2d1ed22a6afeffffff844ce271fb05ce083807cc48dc418d35604f2319946346c92d0705fb6bc21ebd000000006a47304402207a1814eefee55e9ee17653b7122cf6352be8d204ce51ecc9b92efa628295e24202200be4daba99d3a01681d74049eca3f256e523ad9d5bb71718409add4b3961417e4121038eae2a1e8fcf7f43788366ce0c5a39753e14bd561b92121306071abf5f5b34bcfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac0ad5f505000000001976a914e361cafbc05a5596fe4378879230000aa28990b088ac6c000000")
+				case "57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92":
+					return nil, nil
+				}
+				return nil, fmt.Errorf("tx %s not defined for test", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				mp, ok := map[string]*bc.MerkleProof{
+					"3f92e4567741afebdb26296c35c370e274d1b1455faaed2e27af499e6c65e759": nil,
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161": nil,
+					"eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a": nil,
+					"d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc": {
+						Index:  3,
+						TxOrID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+						Target: "67641544b6fe809a0037c9683df0983e973c77a7ed1e8eff3b0f8c9b74cd3472",
+						Nodes: []string{
+							"a6d45ccb68a19c05346d3412f78d154d5834380413cbe740ec3550469b487f6e",
+						},
+					},
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+						Index:  4,
+						TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"*",
+							"*",
+							"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+						},
+					},
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						Index:  6,
+						TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+						Nodes: []string{
+							"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+							"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+							"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+						},
+					},
+					"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92": nil, // The missing tx
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			expErr: errors.New("could not find tx 57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92"),
+		},
+		"error getting tx multiple layers down is handled": {
+			tx: "0200000002bcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4010000006b483045022100d18b422f8cc7c14444aa091fddcb2cf2276c3f7cf496fc186327366f72e8b03802204b6f1b2d2ca44c8a56766f287c951a14302992f55a3063b7fd5c85ec6f2c4f2641210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff6111fe484db08f2a3b6035e21423dc0a8251904951182a9f5b1c3165be79dade010000006b4830450221008a277ad76dfeb69dfdcb15b22ed4539e06d263710e527b676915e90341c19bee02205ae03bbcf4281c2d1470b9e60847f13631c9285557c94df07bc8b447370644784121021f163f44d261868142986872ea90f3155a83775fb9c00b4f3c517e95b30d3b3cfeffffff0200ca9a3b000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac1ed2f505000000001976a914e13aac4a3f01fd3154cc66fc9e285a33bd84d7df88ac6e000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				switch txID {
+				case "deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161":
+					return bt.NewTxFromString("02000000020a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb000000006b483045022100ed01b49050b67634557cf01ad1abcb7271886e65c2edb09c325df3804270ed6402203b86581c0305778e9fb9f4a6b9b1e46a3084e6b85f5676bbdabe31c8128308d5412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff0a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb010000006a4730440220085479effe11ea67ed66622b1b231ec250131cad773bd0b36e79ea81701e7e55022013f9c6ed94ce7099fe807cbf04095cd545b651b6bbe31645ff5ab143b26faf26412102b552f6dff6e03a36ec58d44b71e20d779e92dbe53b52ac08d8236bf87f6ed14dfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac94d3f505000000001976a9141dcd2b24fe8457c775e1a6280bf91d68f48186e988ac6d000000")
+				case "bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84":
+					return bt.NewTxFromString("0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000")
+				case "eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a":
+					return bt.NewTxFromString("0200000003923b8561e34c743ec4e736973f99b4d99d6a6a779d85a29a1f900af435467257010000006b483045022100b3bb6382590f35ce3eeccf3e4e6cc575b2d0f47a5cfb333c77df4ef276cbaa8c022059040d6ce37292f10d7b70feb96d28784d701ca12d0d26adf149400c47e77128412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffbcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4000000006a473044022077eca6d671d07a19c622cf88cb421787e5a11ba3a4a537e00c74be8491966b84022035279d55889a163b53d029e8e65f3430756a59293208cdf371692bc1120276c6412103c2db1771958c949a26ba221dd048dd387f95ea5d761805bbb8c6cd2d1ed22a6afeffffff844ce271fb05ce083807cc48dc418d35604f2319946346c92d0705fb6bc21ebd000000006a47304402207a1814eefee55e9ee17653b7122cf6352be8d204ce51ecc9b92efa628295e24202200be4daba99d3a01681d74049eca3f256e523ad9d5bb71718409add4b3961417e4121038eae2a1e8fcf7f43788366ce0c5a39753e14bd561b92121306071abf5f5b34bcfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac0ad5f505000000001976a914e361cafbc05a5596fe4378879230000aa28990b088ac6c000000")
+				case "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3":
+					return nil, errors.New("close but no cigar")
+				}
+				return nil, fmt.Errorf("tx %s not defined for test", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				mp, ok := map[string]*bc.MerkleProof{
+					"3f92e4567741afebdb26296c35c370e274d1b1455faaed2e27af499e6c65e759": nil,
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161": nil,
+					"eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a": nil,
+					"d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc": {
+						Index:  3,
+						TxOrID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+						Target: "67641544b6fe809a0037c9683df0983e973c77a7ed1e8eff3b0f8c9b74cd3472",
+						Nodes: []string{
+							"a6d45ccb68a19c05346d3412f78d154d5834380413cbe740ec3550469b487f6e",
+						},
+					},
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": nil, // the erroring tx
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						Index:  6,
+						TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+						Nodes: []string{
+							"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+							"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+							"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+						},
+					},
+					"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92": {
+						Index:  3,
+						TxOrID: "57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			expErr: errors.New("failed to get tx 9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3: close but no cigar"),
+		},
+		"error getting merkle proof multiple layers down is handled": {
+			tx: "0200000002bcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4010000006b483045022100d18b422f8cc7c14444aa091fddcb2cf2276c3f7cf496fc186327366f72e8b03802204b6f1b2d2ca44c8a56766f287c951a14302992f55a3063b7fd5c85ec6f2c4f2641210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff6111fe484db08f2a3b6035e21423dc0a8251904951182a9f5b1c3165be79dade010000006b4830450221008a277ad76dfeb69dfdcb15b22ed4539e06d263710e527b676915e90341c19bee02205ae03bbcf4281c2d1470b9e60847f13631c9285557c94df07bc8b447370644784121021f163f44d261868142986872ea90f3155a83775fb9c00b4f3c517e95b30d3b3cfeffffff0200ca9a3b000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac1ed2f505000000001976a914e13aac4a3f01fd3154cc66fc9e285a33bd84d7df88ac6e000000",
+			txFunc: func(ctx context.Context, txID string) (*bt.Tx, error) {
+				switch txID {
+				case "deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161":
+					return bt.NewTxFromString("02000000020a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb000000006b483045022100ed01b49050b67634557cf01ad1abcb7271886e65c2edb09c325df3804270ed6402203b86581c0305778e9fb9f4a6b9b1e46a3084e6b85f5676bbdabe31c8128308d5412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff0a4c8d75e21ee64b498917337dadafc02dc02e05159252a8249a931efae148eb010000006a4730440220085479effe11ea67ed66622b1b231ec250131cad773bd0b36e79ea81701e7e55022013f9c6ed94ce7099fe807cbf04095cd545b651b6bbe31645ff5ab143b26faf26412102b552f6dff6e03a36ec58d44b71e20d779e92dbe53b52ac08d8236bf87f6ed14dfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac94d3f505000000001976a9141dcd2b24fe8457c775e1a6280bf91d68f48186e988ac6d000000")
+				case "bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84":
+					return bt.NewTxFromString("0200000005eb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f000000006b483045022100d49aded8eb72cf3013ff5ab40e72599120eca0efea888746280d139e780e7a8a02202311185979582154595bf54e2df451921350a99fb8dc6b4534070bcceb782e9e412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffeb985192c9a0e72ec42b8b0a6057892f21d1704729e3ae8cfe7db1542788e34f010000006a4730440220183793b2ce4f0038eb9ae1117376a58048405bc842ad3aee0dc5c2edb9210e8e0220706abd0c23ec08d4e88773134e38c138918708e9c3473b864bd5cf7ab057f8e1412102d17ed18d28f653fbaa5c01856086f891447207fc3ea794fcbd8d9ba2ffbd0c68feffffff6c1d183beaa8a17316d9e5d6b16ddc79c726e28e3e2a1819bc974e7cdc5e89c3010000006b483045022100902a1a7ff861fddb44dd61c2ac79f8643a10c727abdc103e7c66696e2e55567a0220380e0f81737daba8d5cc6de763dee708ec42eae9015f31ff42cb6bf6f474b879412102ba8d1245a230a671dff96b97b539901f1a2acf1375cc3b3f15a2d65b05751335fefffffff31d9c18c9a153163ae7b555223573af3abc7f75301d4dbc0b9f172c0302b09e000000006a47304402200fe2ec5cb6209a7a70a115d7a989ec718faa7f107e777e7e8acd1bf13f394d95022019bde30bdf0344354f91152533c1262d2b86bb43d22966735c189f4d1f44bcd6412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffff346ec86bcf97ce29c8d2be62df306e029770c42b4486a35d81e2201aca02af5e000000006a47304402203e04ed6793f42278eb63f63db5bedf518eb58f04b3abb3b1def3206aaf5c056d0220240ea45a398f3ea4eddff6589ba91a90600ac8496f462f27bbf7c455be9df67a41210245189775c1532bedb6b9d929fc539584b07b824a47a5e53af987c92286ba67bafeffffff02f6d7f505000000001976a9142da49f1c71288a4bbf57d0f9aaf352926ef8d66588ac004e7253000000001976a914c29f2b11dbc426d265eb1246463094a06bb0de4988ac67000000")
+				case "eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a":
+					return bt.NewTxFromString("0200000003923b8561e34c743ec4e736973f99b4d99d6a6a779d85a29a1f900af435467257010000006b483045022100b3bb6382590f35ce3eeccf3e4e6cc575b2d0f47a5cfb333c77df4ef276cbaa8c022059040d6ce37292f10d7b70feb96d28784d701ca12d0d26adf149400c47e77128412103a3d76b2838aed38d800b7bc5612add3e90894893ce7489f055f88ec9a6cd22effeffffffbcd68aa8f1c3afcdf1070675c4fa1dfa8c5ae904ee326a61a50ecfe272a785d4000000006a473044022077eca6d671d07a19c622cf88cb421787e5a11ba3a4a537e00c74be8491966b84022035279d55889a163b53d029e8e65f3430756a59293208cdf371692bc1120276c6412103c2db1771958c949a26ba221dd048dd387f95ea5d761805bbb8c6cd2d1ed22a6afeffffff844ce271fb05ce083807cc48dc418d35604f2319946346c92d0705fb6bc21ebd000000006a47304402207a1814eefee55e9ee17653b7122cf6352be8d204ce51ecc9b92efa628295e24202200be4daba99d3a01681d74049eca3f256e523ad9d5bb71718409add4b3961417e4121038eae2a1e8fcf7f43788366ce0c5a39753e14bd561b92121306071abf5f5b34bcfeffffff020027b929000000001976a9147d3722c7bab725e732d57c5db1d2765078aab6c388ac0ad5f505000000001976a914e361cafbc05a5596fe4378879230000aa28990b088ac6c000000")
+				}
+				return nil, fmt.Errorf("tx %s not defined for test", txID)
+			},
+			mpFunc: func(ctx context.Context, txID string) (*bc.MerkleProof, error) {
+				if txID == "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c" {
+					return nil, errors.New("no proof for you")
+				}
+				mp, ok := map[string]*bc.MerkleProof{
+					"3f92e4567741afebdb26296c35c370e274d1b1455faaed2e27af499e6c65e759": nil,
+					"bd1ec26bfb05072dc946639419234f60358d41dc48cc073808ce05fb71e24c84": nil,
+					"deda79be65311c5b9f2a1851499051820adc2314e235603b2a8fb04d48fe1161": nil,
+					"eb48e1fa1e939a24a8529215052ec02dc0afad7d331789494be61ee2758d4c0a": nil,
+					"d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc": {
+						Index:  3,
+						TxOrID: "d485a772e2cf0ea5616a32ee04e95a8cfa1dfac4750607f1cdafc3f1a88ad6bc",
+						Target: "67641544b6fe809a0037c9683df0983e973c77a7ed1e8eff3b0f8c9b74cd3472",
+						Nodes: []string{
+							"a6d45ccb68a19c05346d3412f78d154d5834380413cbe740ec3550469b487f6e",
+						},
+					},
+					"4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb": {
+						Index:  1,
+						TxOrID: "4fe3882754b17dfe8caee3294770d1212f8957600a8b2bc42ee7a0c9925198eb",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"3ba440c59dfa71bad3f8cf8b11c74aa16c3f5cdf95def774c2797cc04075affb",
+							"db5c17b1e18b2ccd42390dfc879d3989010ad14b568fb85545a4cce557eb2675",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c": {
+						Index:  2,
+						TxOrID: "c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+					"9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3": {
+						Index:  4,
+						TxOrID: "9eb002032c179f0bbc4d1d30757fbc3aaf73352255b5e73a1653a1c9189c1df3",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"*",
+							"*",
+							"967b0f389e91648a50f70727bea61bbe7579efee803474780592e3d07a5327f7",
+						},
+					},
+					"5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34": {
+						Index:  6,
+						TxOrID: "5eaf02ca1a20e2815da386442bc47097026e30df62bed2c829ce97cf6bc86e34",
+						Target: "1984451251001b3770f71a6c7beb291a02112d7b93174b58b4be0baae43a0f74",
+						Nodes: []string{
+							"d72e6459a7402c0327c21a7cf8bb2a4e8e006ff0a0699e9f7b65a4db61bcf368",
+							"ce4fdb7ba7a5a70acbe3a76ce0af60c162109a697fd181ddbd7feab8194b226e",
+							"4a259318801977c0970becf54f591b91f426b9a15cc6bb09c960c937e2483837",
+						},
+					},
+					"57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92": {
+						Index:  3,
+						TxOrID: "57724635f40a901f9aa2859d776a6a9dd9b4993f9736e7c43e744ce361853b92",
+						Target: "3df51f7b4aea26ad36c5a6742f8574bcc4abd5f240c7729aef54f8a53f2a500a",
+						Nodes: []string{
+							"c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c",
+							"f2fea4f224492c409ad3bd31845187f5017b4cbceab69f45e922ff1c2d7ac8d5",
+							"4a9ca540c926b5044b1841a622865106618a2613bfe36b36513ce22006983616",
+						},
+					},
+				}[txID]
+				if !ok {
+					return nil, fmt.Errorf("merkle proof for tx %s not defined in test", txID)
+				}
+
+				return mp, nil
+			},
+			expErr: errors.New("failed to get merkle proof for tx c3895edc7c4e97bc19182a3e8ee226c779dc6db1d6e5d91673a1a8ea3b181d6c: no proof for you"),
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			testTx, err := bt.NewTxFromString(test.tx)
+			assert.NoError(t, err)
+
+			mock := &mockTxMerkleGetter{
+				txStoreFunc: test.txFunc,
+				mpStoreFunc: test.mpFunc,
+			}
+
+			c, err := spv.NewCreator(mock, mock)
+			assert.NoError(t, err)
+
+			envelope, err := c.CreateEnvelope(context.TODO(), testTx)
+			if test.expErr == nil {
+				assert.NoError(t, err)
+				assert.NotNil(t, envelope)
+				assert.Equal(t, test.exp, *envelope)
+			} else {
+				assert.Error(t, err)
+				assert.EqualError(t, err, test.expErr.Error())
+			}
+		})
+	}
+}
